refactor: move marker into seperate components
